+++ CHANGELOG.md
@@ -8,11 +8,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [1.0.9] - 2019-04-04
+## [1.0.10] - 2019-04-06
 
 ### Added
 
 - Added view crafting to craft:all
+- Added --extends and --section options (for views)
+- Added --no-views to craft:all command
+    Note: When using craft:all, the --no-create, --no-edit, --no-index and --no-show options are not used
+
+## [1.0.9] - 2019-04-05
+
+### Fixed
+
+- Refactor tests, cleanup 
 
 ## [1.0.8] - 2019-04-04